requirements.txt
altair
pandas
streamlit
psycopg2-binary==2.9.3
